Phase 2 Complete: Reorganize structure (flatten builders, create translators package)
